==> 00. get_reviews.ipynb <==
# Study: Natural Language Processing with Deep Learning
# Dataset: Dead By Daylight Steam Reviews
# Author: Willian Oliveira and Julierme Silva
# Start: 03/04/2023
# Study Motivation: Train a machine to classify products based on user reviews
# Notebook Motivation: Build a dataset with reviews from Dead By Daylight using Steam Reviews API
# Study Status: Finished

# Importing libraries and setting up the environment

import requests
import pandas as pd
from urllib.parse import quote

API_KEY = '<INSERT VALID API KEY>'  
APP_ID = 381210  # Dead By Daylight
NUM_REVIEWS_PER_REQUEST = 100 
LANGUAGE = 'english'
BASE_URL = f"https://store.steampowered.com/appreviews/{APP_ID}?json=1&filter=recent&num_per_page={NUM_REVIEWS_PER_REQUEST}&language={LANGUAGE}&key={API_KEY}"

# Get reviews from Steam API

reviews_list = []
cursor = '*'  # start with cursor set to *

while cursor:
    encoded_cursor = quote(cursor) # encode cursor to be used in URL
    url = f"{BASE_URL}&cursor={encoded_cursor}"
    response = requests.get(url)
    reviews_data = response.json()

    # Verify if the request was successful
    if reviews_data['success'] != 1:
        print("Error in request.")
        break

    for review in reviews_data['reviews']:
        reviews_list.append({
            'review_id': review['recommendationid'],
            'review_text': review['review'],
            'recommended': review['voted_up']
        })
    
    print(f'Reviews colected from Cursor: {cursor}')

    cursor = reviews_data['cursor']  

    if not reviews_data['reviews']:
        print(f'Colected {len(reviews_list)} reviews.')
        print('No more reviews to collect.')
        break



# Save reviews in a DataFrame

reviews_df = pd.DataFrame(reviews_list)
reviews_df.head()

# Save the DataFrame in a CSV file

reviews_df.to_csv(f'data/raw/dbd_{LANGUAGE}_reviews.csv', index=False)
