==> 03. SVM_W2V.ipynb <==
# Study: Natural Language Processing with Deep Learning
# Dataset: Dead By Daylight Steam Reviews
# Author: Willian Oliveira and Julierme Silva
# Start: 10/04/2023
# Study Motivation: Train a machine to classify products based on user reviews
# Notebook Motivation: The purpose of this notebook is to train a Support Vector Machine model to classify the reviews using word2vec.
# Study Status: Finished

# Importing the libraries and setting up the environment

import os
import random
import pandas as pd
import numpy as np
from sklearn.svm import SVC
from gensim.models import Word2Vec
from imblearn.under_sampling import RandomUnderSampler
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.decomposition import PCA
from sklearn.pipeline import Pipeline
from sklearn.model_selection import cross_validate
from sklearn.metrics import (
    accuracy_score,
    precision_score,
    recall_score,
    f1_score,
    roc_auc_score,
    make_scorer,
)
from joblib import dump

SEED = 0


def set_seeds(seed=SEED):
    os.environ["PYTHONHASHSEED"] = str(seed)
    random.seed(seed)
    np.random.seed(seed)


set_seeds()  # Setting seed for reproducible code

# Creating useful functions for this notebook

def review_to_vector(review, model):
    """
    This function takes a preprocessed review text as input and performs the following steps:
    1. Split the review text into words
    2. Retrieve word vectors for each word from the trained Word2Vec model
    3. If no word vectors are found, return a zero vector with the same dimensions as the Word2Vec model's vectors
    4. Otherwise, return the mean of the word vectors as a single vector representation of the review
    
    :param review: str
    :return: review_vector: numpy.ndarray
    """
    words = review.split()
    word_vectors = [model.wv[word] for word in words if word in model.wv.key_to_index]
    
    if len(word_vectors) == 0:
        return np.zeros(model.vector_size)
    else:
        return np.mean(word_vectors, axis=0)

# Importing the dataset

X_train = np.load('data\processed\dbd_proc_train.npz', allow_pickle=True)['X_train']
y_train = np.load('data\processed\dbd_proc_train.npz', allow_pickle=True)['y_train']
X_val = np.load('data\processed\dbd_proc_val.npz', allow_pickle=True)['X_val']
y_val = np.load('data\processed\dbd_proc_val.npz', allow_pickle=True)['y_val']
X_test = np.load('data\processed\dbd_proc_test.npz', allow_pickle=True)['X_test']
y_test = np.load('data\processed\dbd_proc_test.npz', allow_pickle=True)['y_test']


# Verifying the shape of the dataset

print('X_train shape: ', X_train.shape, 'y_train shape: ', y_train.shape)
print('X_val shape: ', X_val.shape, 'y_val shape: ', y_val.shape)
print('X_test shape: ', X_test.shape, 'y_test shape: ', y_test.shape)

# Concatenate train and validation sets since we are using Cross Validation

X_train = np.concatenate((X_train, X_val))
y_train = np.concatenate((y_train, y_val))

# Verifying the first 5 reviews and their labels

for i in range(5):
    print('Review: ', X_train[i])
    print('Label:  ', y_train[i],'\n')


# Balancing the dataset with Random Over Sampling

rus = RandomUnderSampler(random_state=SEED)
X_train_resampled, y_train_resampled = rus.fit_resample(X_train.reshape(-1, 1), y_train)


# verify the shape of the resampled train dataset

print("X_train_resampled shape: ", X_train_resampled.shape, "y_train_resampled shape: ", y_train_resampled.shape)

# verify proportion of positive and negative reviews

print("Positive reviews: ", np.sum(y_train_resampled == 1))
print("Negative reviews: ", np.sum(y_train_resampled == 0))

# Converting X_train_resampled back to a 1D array to be used in the Word2Vec model

X_train_resampled = X_train_resampled.reshape(-1)

# Creating a word2vec class to use in the scikit-learn pipeline


class GensimWord2VecVectorizer(BaseEstimator, TransformerMixin):
    def __init__(
        self, vector_size=100, window=5, min_count=1, workers=4, sg=1, seed=None
    ):
        self.vector_size = vector_size
        self.window = window
        self.min_count = min_count
        self.workers = workers
        self.sg = sg
        self.seed = seed
        self.model = None

    def fit(self, X, y=None):
        self.model = Word2Vec(
            X,
            vector_size=self.vector_size,
            window=self.window,
            min_count=self.min_count,
            workers=self.workers,
            sg=self.sg,
            seed=self.seed,
        )
        return self

    def transform(self, X):
        return np.array(
            [
                np.mean(
                    [self.model.wv[word] for word in words if word in self.model.wv]
                    or [np.zeros(self.vector_size)],
                    axis=0,
                )
                for words in X
            ]
        )

# Training the SVM model with Word2Vec in a pipeline

scoring_metrics = {
    "accuracy": make_scorer(accuracy_score),
    "precision": make_scorer(precision_score),
    "recall": make_scorer(recall_score),
    "f1": make_scorer(f1_score),
    "roc_auc": make_scorer(roc_auc_score),
}

pipeline = Pipeline(
    [
        ("vectorizer", GensimWord2VecVectorizer()),
        ("pca", PCA(n_components=10)),
        ("classifier", SVC(kernel="rbf", # rbf kernel is used due to faster convergence and better performance on tuning tests
                           random_state=SEED,
                           verbose=True,
                           tol=1e-3,
                           max_iter=-1,
                           probability=True)),
    ]
)

cv_results = cross_validate(pipeline,
                            X_train_resampled,
                            y_train_resampled,
                            cv=5,
                            scoring=scoring_metrics)

print("\nCross Validation Results:")
for metric, scores in cv_results.items():
    if "test" in metric:
        print(f"{metric}: {scores.mean():.2f} (+/- {scores.std() * 2:.2f})")

# Using the holdout set to evaluate the final model

pipeline.fit(X_train_resampled, y_train_resampled)
y_pred = pipeline.predict(X_test)

print("\nHoldout Set Results:")
print(f"Accuracy: {accuracy_score(y_test, y_pred):.2f}")
print(f"Precision: {precision_score(y_test, y_pred):.2f}")
print(f"Recall: {recall_score(y_test, y_pred):.2f}")
print(f"F1: {f1_score(y_test, y_pred):.2f}")
print(f"ROC AUC: {roc_auc_score(y_test, y_pred):.2f}")

# Saving the model

dump(pipeline, "models/dbd_reviews_svm_w2v.joblib")
